+++ R/object-usage.R
@@ -2,11 +2,13 @@ object_usage <- function(x) {
   UseMethod("object_usage")
 }
 
-object_usage.default <- function(x) NULL
+object_usage.default <- function(x) {
+  NULL
+}
 
-object_usage.NULL <- function(x) NULL
-
-object_usage.data <- function(x) rd(x$alias)
+object_usage.data <- function(x) {
+  rd(x$alias)
+}
 
 object_usage.function <- function(x) {
   function_usage(x$alias, formals(x$value), identity)
@@ -36,11 +38,6 @@ object_usage.s4method <- function(x) {
   }
   function_usage(x$value@generic, formals(x$value), s4method)
 }
-
-object_usage.s4class <- function(x) NULL
-
-object_usage.rcclass <- function(x) NULL
-
 
 # Function usage ----------------------------------------------------------
 
@@ -90,7 +87,7 @@ usage_args <- function(args) {
 
     text
   }
-  vapply(args, arg_to_text, character(1))
+  map_chr(args, arg_to_text)
 }
 
 args_string <- function(x) {
@@ -102,4 +99,13 @@ args_string <- function(x) {
   arg_names[needs_backtick] <- paste0("`", arg_names[needs_backtick], "`")
 
   paste0(arg_names, sep, x, collapse = ", ")
+}
+
+
+# helpers -----------------------------------------------------------------
+
+# used for testing
+call_to_usage <- function(code, env = pkg_env()) {
+  obj <- call_to_object(!!enexpr(code), env)
+  gsub("\u{A0}", " ", as.character(object_usage(obj)))
 }