+++ R/rd-family.R
@@ -1,3 +1,28 @@
+topics_process_family_prefix <- function(family) {
+
+  default <- paste0("Other ", family, ": ")
+
+  # check for meta (use default prefix when unset)
+  meta <- roxy_meta_get("rd_family_title")
+  if (is.null(meta))
+    return(default)
+
+  # validate meta structure
+  valid <- is.character(meta) || is.list(meta)
+  if (!valid) {
+    message <- "rd_family_title is set, but is not a named list / vector"
+    abort(message)
+  }
+
+  # extract element
+  prefix <- meta[[family]]
+  if (is.null(prefix))
+    return(default)
+
+  prefix
+
+}
+
 topics_process_family <- function(topics) {
   family_index <- invert(topics$simple_values("family"))
   aliases <- topics$simple_values("alias")
@@ -14,12 +39,12 @@ topics_process_family <- function(topics) {
       if (length(others) < 1)
         next
 
-      by_file <- vapply(aliases[others], function(x) {
+      by_file <- map_chr(aliases[others], function(x) {
         paste0("\\code{\\link{", escape(x[1]), "}}")
-      }, FUN.VALUE = character(1))
+      })
       links <- paste(sort_c(by_file), collapse = ", ")
 
-      seealso <- paste("Other ", family, ": ", sep = "")
+      seealso <- topics_process_family_prefix(family)
       out <- strwrap(links, initial = seealso, width = 60, exdent = 2)
 
       topic$add_simple_field("seealso", paste(out, collapse = "\n"))