+++ R/rd.R
@@ -3,11 +3,24 @@ NULL
 
 #' Roclet: make Rd files.
 #'
-#' @family roclets
 #' @template rd
+#' @family roclets
 #' @eval rd_roclet_description()
-#' @seealso `vignette("rd", package = "roxygen2")`
+#' @eval tag_aliases(roclet_tags.roclet_rd)
 #' @export
+#' @examples
+#' #' The length of a string (in characters)
+#' #'
+#' #' @param x String input character vector
+#' #' @return An integer vector the same length as `x`.
+#' #'   `NA` strings have `NA` length.
+#' #' @seealso [nchar()]
+#' #' @export
+#' #' @examples
+#' #' str_length(letters)
+#' #' str_length(c("i", "like", "programming", NA))
+#' str_length <- function(x) {
+#' }
 rd_roclet <- function() {
   roclet("rd")
 }
@@ -60,7 +73,7 @@ roclet_tags.roclet_rd <- function(x) {
     source = tag_markdown,
     template = tag_value,
     templateVar = tag_name_description,
-    title = tag_markdown_restricted,
+    title = tag_markdown,
     usage = tag_value
   )
 }
@@ -71,6 +84,7 @@ roclet_process.roclet_rd <- function(x,
                                      env,
                                      base_path,
                                      global_options = list()) {
+
   # Convert each block into a topic, indexed by filename
   topics <- RoxyTopics$new()
 
@@ -110,7 +124,7 @@ block_to_rd <- function(block, base_path, env, global_options = list()) {
     return()
   }
 
-  name <- block$name %||% object_topic(attr(block, "object"))
+  name <- block$name %||% attr(block, "object")$topic
   if (is.null(name)) {
     block_warning(block, "Missing name")
     return()
@@ -157,7 +171,7 @@ block_to_rd <- function(block, base_path, env, global_options = list()) {
 roclet_output.roclet_rd <- function(x, results, base_path, ..., is_first = FALSE) {
   man <- normalizePath(file.path(base_path, "man"))
 
-  contents <- vapply(results, format, wrap = FALSE, FUN.VALUE = character(1))
+  contents <- map_chr(results, format, wrap = FALSE)
   paths <- file.path(man, names(results))
 
   # Always check for roxygen2 header before overwriting NAMESPACE (#436),
@@ -184,9 +198,7 @@ roclet_output.roclet_rd <- function(x, results, base_path, ..., is_first = FALSE
 roclet_clean.roclet_rd <- function(x, base_path) {
   rd <- dir(file.path(base_path, "man"), full.names = TRUE)
   rd <- rd[!file.info(rd)$isdir]
-  made_by_me <- vapply(rd, made_by_roxygen, logical(1))
-
-  unlink(rd[made_by_me])
+  unlink(purrr::keep(rd, made_by_roxygen))
 }
 
 block_tags <- function(x, tag) {
@@ -274,12 +286,12 @@ topic_add_methods <- function(topic, block) {
   if (is.null(obj$methods)) return()
 
   desc <- lapply(methods, function(x) docstring(x$value@.Data))
-  usage <- vapply(methods, function(x) {
+  usage <- map_chr(methods, function(x) {
     usage <- function_usage(x$value@name, formals(x$value@.Data))
     as.character(wrap_string(usage))
-  }, character(1))
+  })
 
-  has_docs <- !vapply(desc, is.null, logical(1))
+  has_docs <- !map_lgl(desc, is.null)
   desc <- desc[has_docs]
   usage <- usage[has_docs]