+++ R/tag.R
@@ -132,7 +132,7 @@ tag_two_part <- function(first, second, required = TRUE) {
   ## override this behavior
 
   function(x) {
-    if (x$val == "") {
+    if (str_trim(x$val) == "") {
       roxy_tag_warning(x, "requires a value")
     } else if (required && !str_detect(x$val, "[[:space:]]+")) {
       roxy_tag_warning(x, "requires ", first, " and ", second)
@@ -143,7 +143,7 @@ tag_two_part <- function(first, second, required = TRUE) {
 
       x$val <- list(
         pieces[, 1],
-        trim_docstring(full_markdown(pieces[, 2]))
+        trim_docstring(markdown_if_active(pieces[, 2], x))
       )
       names(x$val) <- c(first, second)
       x
@@ -237,13 +237,6 @@ tag_examples <- function(x) {
 #' @export
 #' @rdname roxy_tag
 tag_markdown <- function(x) {
-  x$val <- full_markdown(x$val)
-  tag_value(x)
-}
-
-#' @export
-#' @rdname roxy_tag
-tag_markdown_restricted <- function(x) {
-  x$val <- restricted_markdown(x$val)
+  x$val <- markdown_if_active(x$val, x)
   tag_value(x)
 }