+++ R/topic.R
@@ -32,6 +32,10 @@ RoxyTopic <- R6::R6Class("RoxyTopic", public = list(
     self$fields[[field_name]]
   },
 
+  get_name = function() {
+    self$get_field("name")$values
+  },
+
   inherits_from = function(type) {
     if (!self$has_field("inherit")) {
       return(character())
@@ -39,7 +43,7 @@ RoxyTopic <- R6::R6Class("RoxyTopic", public = list(
 
     inherit <- self$get_field("inherit")
 
-    inherits_field <- vapply(inherit$fields, function(x) type %in% x, logical(1))
+    inherits_field <- map_lgl(inherit$fields, function(x) type %in% x)
     sources <- inherit$source[inherits_field]
 
     if ("NULL" %in% sources)