+++ R/utils.R
@@ -97,33 +97,8 @@ same_contents <- function(path, contents) {
   identical(text_hash, file_hash)
 }
 
-r_files <- function(path) {
-  sort_c(dir(file.path(path, "R"), "\\.[Rr]$", full.names = TRUE))
-}
-
-ignore_files <- function(rfiles, path) {
-  rbuildignore <- file.path(path, ".Rbuildignore")
-  if (!file.exists(rbuildignore))
-    return(rfiles)
-
-  # Strip leading directory and slashes
-  rfiles_relative <- sub(normalizePath(path, winslash = "/"), "", normalizePath(rfiles, winslash = "/"), fixed = TRUE)
-  rfiles_relative <- sub("^[/]*", "", rfiles_relative)
-
-  # Remove any files that match any perl-compatible regexp
-  patterns <- read_lines(rbuildignore)
-  patterns <- patterns[patterns != ""]
-  if (length(patterns) == 0L) {
-    return(rfiles)
-  }
-  matches <- lapply(patterns, grepl, rfiles_relative, perl = TRUE)
-  matches <- Reduce("|", matches)
-  rfiles[!matches]
-}
-
 compact <- function(x) {
-  null <- vapply(x, is.null, logical(1))
-  x[!null]
+  x[!map_lgl(x, is.null)]
 }
 
 block_eval <- function(tag, block, env, tag_name) {
@@ -190,4 +165,14 @@ collapse <- function(key, value, fun, ...) {
 
 cat_line <- function(...) {
   cat(..., "\n", sep = "")
+}
+
+tag_aliases <- function(f) {
+  paste0("@aliases ", paste0("@", names(f()), collapse = " "))
+}
+
+pkg_env <- function() {
+  env <- new.env(parent = globalenv())
+  env$.packageName <- "roxygen2"
+  env
 }