+++ tests/testthat/test-rd-family.R
@@ -95,3 +95,24 @@ test_that("family also included in concepts", {
 
   expect_equal(out$get_field("concept")$values, "a")
 })
+
+test_that("custom family prefixes can be set", {
+
+  owd <- setwd(tempdir())
+  on.exit(setwd(owd), add = TRUE)
+
+  roxy_meta_set("rd_family_title", list(a = "Custom prefix: "))
+  out <- roc_proc_text(rd_roclet(), "
+    #' foo
+    #' @family a
+    foo <- function() {}
+
+    #' bar
+    #' @family a
+    bar <- function() {}
+  ")[[1]]
+
+  seealso <- get_tag(out, "seealso")$values
+  expect_match(seealso, "^Custom prefix:")
+
+})