+++ tests/testthat/test-rd.R
@@ -95,3 +95,87 @@ test_that("documenting NA gives useful error message (#194)", {
     "Missing name"
   )
 })
+
+
+# format ------------------------------------------------------------------
+
+test_that("@format overrides defaults", {
+  out <- roc_proc_text(rd_roclet(), "
+    #' Title
+    #'
+    #' @format abc
+    #'
+    x <- list(a = 1, b = 2)")[[1]]
+
+  expect_equal(get_tag(out, "format")$values, "abc")
+})
+
+test_that("@format NULL suppresses default usage", {
+  out <- roc_proc_text(rd_roclet(), "
+    #' Title
+    #'
+    #' @format NULL
+    #'
+    x <- list(a = 1, b = 2)")[[1]]
+
+  expect_equal(get_tag(out, "format")$values, NULL)
+})
+
+test_that("@format not escaped", {
+  out <- roc_proc_text(rd_roclet(), "
+    #' Title
+    #' @format %
+    #'
+    x <- list(a = 1, b = 2)")[[1]]
+
+  expect_equal(get_tag(out, "format")$values, "%")
+  expect_equal(format(get_tag(out, "format")), "\\format{%}")
+})
+
+
+# UTF-8 -------------------------------------------------------------------
+
+test_that("can generate nonASCII document", {
+  test_pkg <- temp_copy_pkg(test_path('testNonASCII'))
+  on.exit(unlink(test_pkg, recursive = TRUE), add = TRUE)
+
+  expect_output(roxygenise(test_pkg, roclets = "rd"), "printChineseMsg[.]Rd")
+
+  rd_path <- file.path(test_pkg, "man", "printChineseMsg.Rd")
+  expect_true(file.exists(rd_path))
+  rd <- read_lines(rd_path)
+
+  expect_true(any(grepl("\u6211\u7231\u4e2d\u6587", rd)))
+  expect_true(any(grepl("\u4e2d\u6587\u6ce8\u91ca", rd)))
+
+  # Shouldn't change again
+  expect_output(roxygenise(test_pkg, roclets = "rd"), NA)
+})
+
+
+test_that("unicode escapes are ok", {
+  test_pkg <- temp_copy_pkg(test_path('testUtf8Escape'))
+  on.exit(unlink(test_pkg, recursive = TRUE), add = TRUE)
+
+  expect_output(roxygenise(test_pkg, roclets = "rd"), "a[.]Rd")
+
+  rd_path <- file.path(test_pkg, "man", "a.Rd")
+  expect_true(file.exists(rd_path))
+  rd <- read_lines(rd_path)
+
+  expect_true(any(grepl("7\u00b0C", rd)))
+
+  # Shouldn't change again
+  expect_output(roxygenise(test_pkg, roclets = "rd"), NA)
+})
+
+test_that("write_lines writes unix-style line endings.", {
+  path <- test_path("escapes.Rd")
+  temp_filename <- tempfile()
+  old_binary <- readBin(path, "raw", n = file.info(path)$size)
+  old_text <- read_lines(path)
+  write_lines(old_text, temp_filename)
+  on.exit(unlink(temp_filename), add = TRUE)
+  new_binary <- readBin(temp_filename, "raw", n = file.info(temp_filename)$size)
+  expect_identical(new_binary, old_binary)
+})